refactor(gateway): extract the protocol and its runtime into packages/gateway
The protocol, the host's server, the client, the transports, the node
registry, the shutdown coordinator, and the socket binding were spread
across @sidecar/runtime-contracts, @sidecar/runtime, and the desktop's own
main/gateway. They are one boundary, so they are one package: @sidecar/gateway
is the API contract and its runtime, the desktop reaches it as an operator
client, and a cloud process binds the same server over the same protocol.

runtime-contracts drops protocol.ts rather than re-exporting it, which would
close a cycle. The reply-grant states go with the ledger that owns them, in
@sidecar/brain's delivery.ts, since nothing about them ever travels on the
wire. The two socket halves become one websocket.ts behind its own door,
because it reaches ws and node:http and a bundle that only wants the
vocabulary must not resolve them; the text transport moves to ./testing,
which is the only thing that composes it.
