+++ packages/gateway/package.json
@@ -1,10 +1,13 @@
 {
-  "name": "@sidecar/runtime",
+  "name": "@sidecar/gateway",
   "version": "0.5.0",
   "private": true,
   "type": "module",
   "exports": {
-    ".": "./src/index.js"
+    ".": "./src/index.js",
+    "./protocol": "./src/protocol.js",
+    "./testing": "./src/testing.js",
+    "./websocket": "./src/websocket.js"
   },
   "types": "./src/index.ts",
   "scripts": {
@@ -14,10 +17,11 @@
   "dependencies": {
     "@sidecar/runtime-contracts": "workspace:*",
     "@sidecar/wire": "workspace:*",
-    "croner": "10.0.1"
+    "ws": "8.21.3"
   },
   "devDependencies": {
     "@types/node": "26.2.0",
+    "@types/ws": "8.18.1",
     "tsx": "4.23.12",
     "typescript": "7.0.2"
   }
