+++ packages/gateway/src/attachment.ts
