+++ packages/runtime-contracts/src/index.ts
@@ -104,53 +104,6 @@ export {
   threadSessionKey,
 } from "./identifiers.js";
 export {
-  DELIVERY_STATE,
-  type DeliveryState,
-  GATEWAY_CLIENT_ROLE,
-  GATEWAY_ERROR,
-  GATEWAY_EVENT,
-  GATEWAY_HANDSHAKE_HEADER,
-  GATEWAY_HANDSHAKE_REFUSAL,
-  GATEWAY_METHOD,
-  GATEWAY_PROTOCOL_VERSION,
-  GATEWAY_RECONNECT_KIND,
-  type GatewayClientIdentity,
-  type GatewayClientRole,
-  type GatewayError,
-  type GatewayErrorCode,
-  type GatewayEvent,
-  type GatewayEventKind,
-  type GatewayExpectedRevision,
-  type GatewayHandshakeRefusal,
-  type GatewayMethod,
-  type GatewayReconnectAnswer,
-  type GatewayReconnectKind,
-  type GatewayRequest,
-  type GatewayResponse,
-  type GatewayRevision,
-  gatewayEventFromWire,
-  gatewayReconnectAnswerFromWire,
-  gatewayRequestFromWire,
-  gatewayResponseFromWire,
-  isGatewayErrorCode,
-  isGatewayEventKind,
-  isGatewayMethod,
-  isMutatingGatewayMethod,
-  isTerminalDeliveryState,
-  NODE_CAPABILITY_STATUS,
-  type NodeCapabilityResult,
-  type NodeCapabilityStatus,
-  type NodeInvocation,
-  type NodeInvocationAnswer,
-  nodeCapabilityResultFromWire,
-  nodeCapabilityResultToWire,
-  nodeInvocationAnswerFromWire,
-  nodeInvocationAnswerToWire,
-  nodeInvocationFromWire,
-  nodeInvocationToWire,
-  TERMINAL_DELIVERY_STATES,
-} from "./protocol.js";
-export {
   ARCHIVE_ENCODING,
   ARCHIVE_REASON,
   type ArchiveEncoding,
