--- packages/runtime/src/gateway/local-transport.ts
@@ -1,256 +1,0 @@
-import type { IncomingMessage } from "node:http";
-import {
-  GATEWAY_ERROR,
-  GATEWAY_HANDSHAKE_HEADER,
-  GATEWAY_HANDSHAKE_REFUSAL,
-  GATEWAY_PROTOCOL_VERSION,
-  type GatewayClientIdentity,
-  type GatewayEvent,
-  type GatewayHandshakeRefusal,
-  type GatewayRequest,
-  type GatewayResponse,
-  gatewayEventFromWire,
-  gatewayResponseFromWire,
-  type NodeInvocation,
-  nodeInvocationAnswerToWire,
-  nodeInvocationFromWire,
-} from "@sidecar/runtime-contracts";
-import { isRecord, type UnparsedWireValue, type WireValue } from "@sidecar/wire";
-import { WebSocket } from "ws";
-import {
-  InvocationMemory,
-  NODE_INVOCATION_REFUSAL,
-  type NodeInvocationHandler,
-  unavailableInvocation,
-} from "./invocations.js";
-import { GATEWAY_FRAME, GATEWAY_REFUSAL_HEADER } from "./local-host.js";
-import type { GatewayEventSink, GatewayTransport } from "./transport.js";
-
-/**
- * The client end of the socket. The credential rides on the upgrade
- * request's authorization header and is dropped from memory once the
- * handshake ends; the address the socket opens names only the host. A
- * request that is in flight when the socket closes answers the typed
- * disconnected error rather than hanging.
- */
-export interface WebSocketGatewayConnectOptions {
-  url: string;
-  /** What authenticates this client, in the header its host's own `authenticate` reads. */
-  headers?: Readonly<Record<string, string>>;
-  client: GatewayClientIdentity;
-  /** How long the handshake may take before the attempt is unreachable. */
-  timeoutMs?: number;
-}
-
-export const WEB_SOCKET_GATEWAY_CONNECT_DEFAULTS = {
-  TIMEOUT_MS: 5_000,
-} as const;
-
-/** A handshake that did not end in a connection: the refusal the host named, or no host at all. */
-export const GATEWAY_UNREACHABLE = "unreachable";
-
-export type GatewayConnectResult =
-  | { ok: true; connection: WebSocketGatewayConnection }
-  | { ok: false; failure: GatewayHandshakeRefusal | typeof GATEWAY_UNREACHABLE };
-
-function isRefusal(value: string | undefined): value is GatewayHandshakeRefusal {
-  return Object.values(GATEWAY_HANDSHAKE_REFUSAL).some((held) => held === value);
-}
-
-function requestToWire(request: GatewayRequest): WireValue {
-  return {
-    protocolVersion: request.protocolVersion,
-    id: request.id,
-    method: request.method,
-    params: request.params,
-    ...(request.idempotencyKey !== undefined
-      ? { idempotencyKey: request.idempotencyKey }
-      : undefined),
-    ...(request.expectedRevision !== undefined
-      ? { expectedRevision: { ...request.expectedRevision } }
-      : undefined),
-  };
-}
-
-function disconnected(id: string): GatewayResponse {
-  return {
-    id,
-    ok: false,
-    error: { code: GATEWAY_ERROR.DISCONNECTED, message: "the Gateway socket is closed" },
-    revision: { configuration: 0, sequence: 0 },
-  };
-}
-
-export class WebSocketGatewayConnection implements GatewayTransport {
-  readonly #socket: WebSocket;
-  readonly #pending = new Map<string, (response: GatewayResponse) => void>();
-  readonly #sinks = new Set<GatewayEventSink>();
-  readonly #closedListeners = new Set<() => void>();
-  #memory: InvocationMemory | undefined;
-  #open = true;
-
-  constructor(socket: WebSocket) {
-    this.#socket = socket;
-    socket.on("message", (data, isBinary) => {
-      if (isBinary) return;
-      this.#take(data.toString());
-    });
-    socket.on("close", () => this.#closed());
-    socket.on("error", () => this.#closed());
-  }
-
-  request(request: GatewayRequest): Promise<GatewayResponse> {
-    if (!this.#open || this.#socket.readyState !== WebSocket.OPEN) {
-      return Promise.resolve(disconnected(request.id));
-    }
-    return new Promise((resolve) => {
-      this.#pending.set(request.id, resolve);
-      this.#socket.send(
-        JSON.stringify({
-          kind: GATEWAY_FRAME.REQUEST,
-          envelope: requestToWire(request),
-        }),
-        (error) => {
-          if (!error) return;
-          this.#pending.delete(request.id);
-          resolve(disconnected(request.id));
-        },
-      );
-    });
-  }
-
-  events(sink: GatewayEventSink): () => void {
-    this.#sinks.add(sink);
-    return () => {
-      this.#sinks.delete(sink);
-    };
-  }
-
-  connected(): boolean {
-    return this.#open && this.#socket.readyState === WebSocket.OPEN;
-  }
-
-  /**
-   * Serves the host's invocations arriving on this socket. Each is deduped
-   * by id before the handler runs, and answered on this socket alone; an
-   * invocation arriving while no handler is served is answered unavailable,
-   * so the host never waits on a node that is not there.
-   */
-  serveInvocations(handler: NodeInvocationHandler): () => void {
-    const memory = new InvocationMemory(handler);
-    this.#memory = memory;
-    return () => {
-      if (this.#memory === memory) this.#memory = undefined;
-    };
-  }
-
-  onClosed(listener: () => void): () => void {
-    this.#closedListeners.add(listener);
-    return () => {
-      this.#closedListeners.delete(listener);
-    };
-  }
-
-  close(): void {
-    if (!this.#open) return;
-    this.#socket.close(1000, "the client is leaving");
-    this.#closed();
-  }
-
-  #take(text: string): void {
-    let value: UnparsedWireValue;
-    try {
-      // SAFETY: JSON.parse returns a wire value; the reader is the validation.
-      value = JSON.parse(text) as UnparsedWireValue;
-    } catch {
-      return;
-    }
-    if (!isRecord(value) || !isRecord(value.envelope)) return;
-    const { kind, envelope } = value;
-    if (kind === GATEWAY_FRAME.RESPONSE) {
-      const response = gatewayResponseFromWire(envelope);
-      if (!response) return;
-      const resolve = this.#pending.get(response.id);
-      if (!resolve) return;
-      this.#pending.delete(response.id);
-      resolve(response);
-      return;
-    }
-    if (kind === GATEWAY_FRAME.EVENT) {
-      const event: GatewayEvent | undefined = gatewayEventFromWire(envelope);
-      if (!event) return;
-      for (const sink of [...this.#sinks]) sink(event);
-      return;
-    }
-    if (kind === GATEWAY_FRAME.INVOCATION) {
-      const invocation = nodeInvocationFromWire(envelope);
-      if (!invocation) return;
-      void this.#answer(invocation);
-    }
-  }
-
-  async #answer(invocation: NodeInvocation): Promise<void> {
-    const memory = this.#memory;
-    const answer = memory
-      ? await memory.take(invocation)
-      : {
-          invocationId: invocation.invocationId,
-          result: unavailableInvocation(invocation, NODE_INVOCATION_REFUSAL.NOT_SERVING),
-        };
-    if (!this.connected()) return;
-    this.#socket.send(
-      JSON.stringify({
-        kind: GATEWAY_FRAME.ANSWER,
-        envelope: nodeInvocationAnswerToWire(answer),
-      }),
-    );
-  }
-
-  #closed(): void {
-    if (!this.#open) return;
-    this.#open = false;
-    for (const [id, resolve] of this.#pending) resolve(disconnected(id));
-    this.#pending.clear();
-    for (const listener of [...this.#closedListeners]) listener();
-    this.#closedListeners.clear();
-    this.#sinks.clear();
-  }
-}
-
-export function connectWebSocketGateway(
-  options: WebSocketGatewayConnectOptions,
-): Promise<GatewayConnectResult> {
-  const { url, client } = options;
-  const timeoutMs = options.timeoutMs ?? WEB_SOCKET_GATEWAY_CONNECT_DEFAULTS.TIMEOUT_MS;
-  return new Promise((resolve) => {
-    const socket = new WebSocket(url, {
-      headers: {
-        ...options.headers,
-        [GATEWAY_HANDSHAKE_HEADER.PROTOCOL_VERSION]: String(GATEWAY_PROTOCOL_VERSION),
-        [GATEWAY_HANDSHAKE_HEADER.CLIENT_ID]: client.clientId,
-        [GATEWAY_HANDSHAKE_HEADER.CLIENT_ROLE]: client.role,
-      },
-      handshakeTimeout: timeoutMs,
-      perMessageDeflate: false,
-    });
-    let settled = false;
-    const settle = (result: GatewayConnectResult) => {
-      if (settled) return;
-      settled = true;
-      resolve(result);
-    };
-    socket.once("unexpected-response", (_request, response: IncomingMessage) => {
-      const refusal = response.headers[GATEWAY_REFUSAL_HEADER];
-      const named = Array.isArray(refusal) ? refusal[0] : refusal;
-      response.resume();
-      socket.terminate();
-      settle({ ok: false, failure: isRefusal(named) ? named : GATEWAY_UNREACHABLE });
-    });
-    socket.once("error", () => {
-      settle({ ok: false, failure: GATEWAY_UNREACHABLE });
-    });
-    socket.once("open", () => {
-      settle({ ok: true, connection: new WebSocketGatewayConnection(socket) });
-    });
-  });
-}
