+++ packages/runtime/src/index.ts
@@ -45,69 +45,6 @@ export {
   scheduledJobFromWire,
   validCronExpression,
 } from "./cron.js";
-export {
-  type GatewayCallOptions,
-  type GatewayCallResult,
-  GatewayClient,
-  type GatewayClientEventListener,
-  type GatewayClientOptions,
-  helloSequence,
-} from "./gateway/client.js";
-export {
-  type DeliveryClaim,
-  type DeliveryClaimContext,
-  DeliveryLedger,
-  type DeliveryLedgerOptions,
-  type DeliveryOffer,
-  type DeliveryRecord,
-  deliveryRecordToWire,
-} from "./gateway/delivery.js";
-export {
-  INVOCATION_MEMORY_DEFAULTS,
-  InvocationMemory,
-  NODE_INVOCATION_REFUSAL,
-  type NodeInvocationHandler,
-  PendingInvocations,
-  unavailableInvocation,
-  unknownInvocation,
-} from "./gateway/invocations.js";
-export {
-  type NodeCapabilityHandler,
-  type NodeRegistration,
-  NodeRegistry,
-  type NodeRegistryListener,
-  type NodeSnapshot,
-  nodeSnapshotToWire,
-  type RemoteNodeInvoker,
-  type RemoteNodeRegistration,
-} from "./gateway/nodes.js";
-export {
-  eventToWire,
-  GATEWAY_SERVER_DEFAULTS,
-  type GatewayEventListener,
-  type GatewayMethodContext,
-  type GatewayMethodHandler,
-  type GatewayMethodOutcome,
-  type GatewayMethodTable,
-  GatewayServer,
-  type GatewayServerOptions,
-  gatewayError,
-  gatewayOk,
-} from "./gateway/server.js";
-export {
-  GATEWAY_SHUTDOWN_DEFAULTS,
-  type GatewayShutdownOptions,
-  type GatewayShutdownReport,
-  type GatewayShutdownSteps,
-  shutdownGateway,
-} from "./gateway/shutdown.js";
-export {
-  type GatewayEventSink,
-  type GatewayTransport,
-  InProcessTransport,
-  TextLoopbackTransport,
-  type TextLoopbackTransportOptions,
-} from "./gateway/transport.js";
 export { HEARTBEAT_DEFAULTS, heartbeatJob } from "./heartbeat.js";
 export {
   agentLaneWidth,
